calendar-events/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-functions
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sendgrid
openai
python-dotenv                                                                                                                                                                                                            
